backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
python-multipart==0.0.6
pydantic==2.5.2
google-api-python-client==2.108.0
watchdog==3.0.0 